enh: adding new bills services
