+++ src/application/bill/bill.module.ts
@@ -5,11 +5,14 @@ import { Bill } from './bill.entity'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { BankModule } from '../bank/bank.module'
 import { CreditCardModule } from '../credit-card/credit-card.module'
+import { UpdateBillService } from './services/update-bill.service'
+import { CreateBillService } from './services/create-bill.service'
+import { GetBillService } from './services/get-bill.service'
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Bill]), BankModule, CreditCardModule],
+  imports: [TypeOrmModule.forFeature([Bill]), BillModule, BankModule, CreditCardModule],
   controllers: [BillController],
-  providers: [BillService],
+  providers: [UpdateBillService, CreateBillService, GetBillService, BillService],
   exports: [TypeOrmModule]
 })
 export class BillModule {}
